+++ KageIDE/KageIDE/Forms/SymbolForm.cs
@@ -43,6 +43,16 @@ namespace KagaIDE.Forms
                 this.funListBox.Items.Add(s);
             }
             this.funListBox.SelectedIndex = 0;
+            // 加载全局变量
+            List<string> globalVector = core.getGlobalVar();
+            this.globalvarDataGridView.Rows.Clear();
+            for (int i = 0; i < globalVector.Count; i++)
+            {
+                this.globalvarDataGridView.Rows.Add();
+                string[] splitItem = globalVector[i].Split('@');
+                this.globalvarDataGridView.Rows[i].Cells[0].Value = splitItem[0];
+                this.globalvarDataGridView.Rows[i].Cells[1].Value = splitItem[1];
+            }
         }
 
         private void funListBox_SelectedIndexChanged(object sender, EventArgs e)
@@ -78,16 +88,31 @@ namespace KagaIDE.Forms
                 MessageBox.Show("主函数main不可以被删除", "错误", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 return;
             }
-            // 否则，调用控制器删除方法
-            core.deleteFunction((string)this.funListBox.SelectedItem);
-            // 刷新前台
-            ((MainForm)this.Owner).functionListBox.Items.Remove((string)this.funListBox.SelectedItem);
-            ((MainForm)this.Owner).closeTabCard((string)this.funListBox.SelectedItem);
-            ((MainForm)this.Owner).tabControl1.SelectedTab = ((MainForm)this.Owner).tabControl1.TabPages[0];
-            Consta.refreshMutex = true;
-            this.funListBox.Items.RemoveAt(this.funListBox.SelectedIndex);
-            Consta.refreshMutex = false;
-            this.funListBox.SelectedIndex = 0;
+            // 提示警告
+            DialogResult dr = MessageBox.Show(
+                String.Format("删除函数是不可撤销的操作！{0}真的要继续吗？", Environment.NewLine),
+                "警告", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning, MessageBoxDefaultButton.Button2);
+            if (dr == System.Windows.Forms.DialogResult.OK)
+            {
+                System.Threading.Thread.Sleep(1000);
+                dr = MessageBox.Show(
+                String.Format("再次警告！{0}删除函数是不可撤销的操作！{1}确认真的要继续吗？", Environment.NewLine, Environment.NewLine),
+                "警告", MessageBoxButtons.YesNoCancel, MessageBoxIcon.Warning, MessageBoxDefaultButton.Button3);
+                if (dr == System.Windows.Forms.DialogResult.Yes)
+                {
+                    // 调用控制器删除方法
+                    core.deleteFunction((string)this.funListBox.SelectedItem);
+                    // 刷新前台
+                    ((MainForm)this.Owner).functionListBox.Items.Remove((string)this.funListBox.SelectedItem);
+                    ((MainForm)this.Owner).closeTabCard((string)this.funListBox.SelectedItem);
+                    ((MainForm)this.Owner).tabControl1.SelectedTab = ((MainForm)this.Owner).tabControl1.TabPages[0];
+                    Consta.refreshMutex = true;
+                    this.funListBox.Items.RemoveAt(this.funListBox.SelectedIndex);
+                    Consta.refreshMutex = false;
+                    this.funListBox.SelectedIndex = 0;
+                }
+            }
+            return;
         }
 
         // 添加按钮
@@ -95,6 +120,97 @@ namespace KagaIDE.Forms
         {
             FunctionForm addFunForm = new FunctionForm("从管理器新建函数");
             addFunForm.ShowDialog(this);
+        }
+
+        private void globalvarDataGridView_EditingControlShowing(object sender, DataGridViewEditingControlShowingEventArgs e)
+        {
+            if (this.globalvarDataGridView.CurrentCell.ColumnIndex == 1)
+            {
+                ComboBox cb = e.Control as ComboBox;
+                if (cb != null)
+                {
+                    cb.SelectedIndex = 0;
+                }
+            }
+        }
+
+        private void globalvarDataGridView_CellMouseClick(object sender, DataGridViewCellMouseEventArgs e)
+        {
+            try
+            {
+                // 单元格不存在时
+                if (e.RowIndex == -1 || e.ColumnIndex == -1)
+                {
+                    return;
+                }
+                // 左键时为类型赋初值
+                if (e.Button == System.Windows.Forms.MouseButtons.Left)
+                {
+                    if ((this.globalvarDataGridView.Rows[e.RowIndex].Cells[0].Value) == null)
+                    {
+                        this.globalvarDataGridView.Rows[e.RowIndex].Cells[1].Value = Consta.basicType[0];
+                    }
+                }
+                // 右键删除变量
+                else if (e.Button == System.Windows.Forms.MouseButtons.Right)
+                {
+                    // 边界要减一，因为最后一行是没有提交的
+                    if (e.RowIndex != this.globalvarDataGridView.Rows.Count - 1)
+                    {
+                        this.globalvarDataGridView.Rows.RemoveAt(e.RowIndex);
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                throw ex;
+            }
+        }
+
+        // 全局变量保存
+        private void button1_Click(object sender, EventArgs e)
+        {
+            // 处理参数列表
+            List<string> argvList = new List<string>();
+            int nrows = this.globalvarDataGridView.Rows.Count - 1;
+            for (int i = 0; i < nrows; i++)
+            {
+                string varname = this.globalvarDataGridView.Rows[i].Cells[0].Value.ToString();
+                // 类型不能为空
+                if (this.globalvarDataGridView.Rows[i].Cells[1].Value == null)
+                {
+                    MessageBox.Show(String.Format("变量 {0} 类型不合法", varname), "错误", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+                string vartype = this.globalvarDataGridView.Rows[i].Cells[1].Value.ToString();
+                // 符号合法性
+                if (Consta.IsStdCSymbol(varname) == false)
+                {
+                    MessageBox.Show(String.Format("变量 {0} 命名不合法", varname), "错误", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+                // 符号唯一性
+                if (argvList.Find((x) => x.Split('@')[0] == varname) != null)
+                {
+                    MessageBox.Show(String.Format("变量名 {0} 重复", varname), "错误", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+                // 类型合法性
+                if (Consta.basicType.Find((x) => x == vartype) == null)
+                {
+                    MessageBox.Show(String.Format("变量 {0} 类型不合法", varname), "错误", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+                argvList.Add(String.Format("{0}@{1}", varname, vartype));
+            }
+            // 把更改提交给后台
+            core.setNewGlobalVar(argvList);
+            // 刷新主窗体
+            ((MainForm)this.Owner).globalvarListBox.Items.Clear();
+            foreach (string s in core.getGlobalVar())
+            {
+                ((MainForm)this.Owner).globalvarListBox.Items.Add((string)s.Replace("@", " @ "));
+            }
         }
     }
 }
